requirements.txt
Flask==3.0.3
Pillow==11.0.0
google-cloud-vision==3.8.0
google-cloud-storage==2.18.2
gunicorn==23.0.0